WIP breaking out recommendation list component.
