+++ src/components/recommendation-list/index.js
@@ -1,0 +1,3 @@
+import RecommendationList from './recommendation-list'
+
+export default RecommendationList
