+++ src/components/recommendation-list/recommendation-list.scss
@@ -1,0 +1,51 @@
+@import "../../styles/variables.scss";
+@import "../../styles/mixins.scss";
+
+recommendation-list {
+  .recommendation {
+    width: 20rem;
+    display: inline-block;
+  }
+
+  .list-box {
+    border: 1px solid $light-grey;
+    max-height: 100%;
+    height: 60%;
+    overflow-y: auto;
+    border: 1px solid black;
+    display: inline-block;
+    width: 23rem;
+    padding: 0.5em;
+
+    &::-webkit-scrollbar {
+        -webkit-appearance: none;
+    }
+
+    &::-webkit-scrollbar:vertical {
+        width: 11px;
+    }
+
+    &::-webkit-scrollbar-thumb {
+        border-radius: 8px;
+        border: 2px solid white;
+        background-color: $light-blue;
+    }
+  }
+
+  .copy-table-name-icon {
+    vertical-align: bottom;
+    color: $light-blue;
+
+    &:hover {
+      color: $almost-black;
+    }
+
+    &:active {
+      color: $medium-grey;
+    }
+  }
+
+  a {
+    @include link();
+  }
+}
